refactored out components of app page into seperate components, extracted fetch wrappers to thier own functions and moved and renamed CSS styles file
