+++ packages/registrar_demo_app/src/utilities/fetch-json.js
@@ -1,0 +1,11 @@
+const fetchJson = async (url) => {
+  try {
+    const response = await fetch(url);
+    const json = await response.json();
+    return json;
+  } catch (error) {
+    console.error(`Error fetching from url ${url}:`, error);
+  }
+}
+
+export default fetchJson;
